Adding explicit optional attribute to indirection requests
Previously, any option that had a setter was treated
as an attribute, but now we're specifying the list of
attributes settable via options.

We also have a to_hash method that will take all of the
options and all of those attributes and join them
back into a hash.  This method is used by the REST Handler
module, since it uses the indirection request internally.

Signed-off-by: Luke Kanies <luke@madstop.com>