Refactored queue registry module to use instance loader behavior
The initial implementation of queue client library management with
Puppet::Util::Queue had custom type/class registry functionality.
Stripped this and replaced it with use of the
Puppet::Util::InstanceLoader module instead.  This required some
significant revision to the unit tests, and some additional registry
functionality on top of the instance loader behavior (as the
instance loader/autoloader stuff does not have the final piece of smarts
to map symbolic type names to specific classes).

Note that, apparently in contrast to how termini do it, this only loads
the client plugin requested through usage of the Puppet::Util::Queue
client/client_class methods.  It will not load all plugin types available,
only the one you try to use.

TO-DO: revise the documentation, particularly for the queue client
library registry (Puppet::Util::Queue) in order to reflect this new
design.