Restored tests for queue client class registration, as well as default behavior of determining queue client type symbols from the client class name.
Tests were adjusted to be a little higher-level (easier for creating test classes,
for instance), and to work with changes in assumptions about how registered clients
are organized.

Added check to Puppet::Queue::Client.queue_type_to_class to throw an exception if
the type provided doesn't map to a class.